format and lint fixes

+++ blockchain/ethereum_gateway.go
@@ -220,7 +220,7 @@ func (g *EthereumGateway) SetContractCloseOut(fromAddress string, contractAddres
 	g.log.Debugf("closeout type: %v", closeoutType)
 
 	//TODO: retry if price is too low
-	tx, err := instance.SetContractCloseOut(options, big.NewInt(int64(closeoutType)))
+	tx, err := instance.SetContractCloseOut(options, big.NewInt(closeoutType))
 	// g.pendingNonce.Unlock()
 	if err != nil {
 		g.log.Errorf("cannot close transaction: %s tx: %s fromAddr: %s contractAddr: %s", err, tx, fromAddress, contractAddress)
